refactor: Rename epic-win files and paths to epicwin for consistency
- Rename directory: epic-win-section → epicwin-section
- Rename files: epic-win-*.tsx → epicwin-*.tsx
- Rename action file: epic-win.ts → epicwin.ts
- Update import paths and URL routes accordingly
- Maintain database schema naming (epicwins table unchanged)
